20201208 Authority check when article reply delete

+++ at/src/main/java/com/sbs/cwy/at/controller/ArticleController.java
@@ -57,7 +57,7 @@ public class ArticleController {
 		String redirectUrl = (String) param.get("redirectUrl");
 		redirectUrl = redirectUrl.replace("#id", newArticleId + "");
 
-		return "redirect:" + redirectUrl;
+		return "common/redirect";
 	}
 
 	@RequestMapping("/usr/article/doWriteReplyAjax")
@@ -86,7 +86,13 @@ public class ArticleController {
 
 	@RequestMapping("/usr/article/doDeleteReplyAjax")
 	@ResponseBody
-	public ResultData doDeleteReplyAjax(int id) {
+	public ResultData doDeleteReplyAjax(int id, HttpServletRequest req) {
+		Member loginedMember = (Member) req.getAttribute("loginedMember");
+		ArticleReply articleReply = articleService.getForPrintArticleReplyById(id);
+
+		if (articleService.actorCanDelete(loginedMember, articleReply) == false) {
+			return new ResultData("F-1", String.format("%d번 댓글을 삭제할 권한이 없습니다.", id));
+		}
 		articleService.deleteReply(id);
 
 		return new ResultData("S-1", String.format("%d번 댓글을 삭제하였습니다.", id));
